+++ lib/util/deep-dots.ts
@@ -3,7 +3,7 @@ export default function (source: any, path: string): any {
   let dots = path.split('.');
   let target = source;
   while (dots.length > 0) {
-    target = source[dots[0]];
+    target = target[dots[0]];
     dots = dots.slice(1);
   }
   return target;