+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 1.1.3
+* Bug fix with deepDots not picking up nested properties
+
 ## 1.1.2
-* Exposing utilities.
+* Exposing utilities
 
 ## 1.1.0
 * Linted code which changed a few class names
@@ -11,4 +14,4 @@
 * Various fixes from real-world use
 
 ## 1.0.0
-First working version.
+First working version